feat(cli): land v0.80.3 CLI module split + stable renderer + cards
HEAD was inconsistent: committed tests imported code-agent-loop.js and other
modules that were never committed, and committed code.ts was still 2199 lines.
Codex went offline mid-release with this work as uncommitted WIP; landing it so
v0.80.3 can be cut from a reproducible commit.

- Split commands/code.ts (2199 -> 884) into code-agent-loop / code-resume /
  code-context / code-runtime-frames / code-tool-protocol / code-tool-render.
- Split worker-run external-agent logic into worker-external-events.
- Stable Apple Terminal renderer (no Ink/raw/fullscreen): low-noise shimmer +
  colored tool/plan cards + static footer/TPS (brain-render, terminal-spinner,
  interactive-line) — IME-safe.
- shared/fleet-events.ts contract update used by the worker adapters.

Working tree was green (cli typecheck 0, build OK, 417 tests pass, largest file
892 < 900). No new 900+ monolith.
