+++ cli/src/code-agent-loop.ts
@@ -1,0 +1,477 @@
+import { streamBrainChat, type BrainStreamEvent, type ChatMessage } from "./brain-client.js";
+import { formatBrainErrorForHuman, renderBrainEventForHuman, summarizeUsage, type HumanBrainRenderState } from "./brain-render.js";
+import { buildCodePrompt, type CodeContext } from "./code-context.js";
+import { buildCodeRuntimeFrames } from "./code-runtime-frames.js";
+import { assistantToolCallsForMessages, codeToolDefinitions, createStreamingToolCallAccumulator, parseCodeToolRequests, toolRequestFingerprint, toolRequestsFromCollectedCalls, type CodeToolRequest, type CollectedToolCall } from "./code-tool-protocol.js";
+import { formatDangerousToolPreview, isDangerousClientTool, redactToolArgs, renderClientToolResult, renderClientToolStart, resolveToolApproval, ToolApprovalRequiredError, type ToolApprovalRequest } from "./code-tool-render.js";
+import { buildCodeContextMessages } from "./context-layers.js";
+import { t } from "./i18n.js";
+import { nowIso, writeJsonLine } from "./jsonl.js";
+import { buildImagesContentParts } from "./media.js";
+import { normalizePlanItems, renderPlanItems, type CodePlanItem } from "./plan-tool.js";
+import type { CliProviderProfile } from "./provider-profile.js";
+import { parseReasoningOptions, shouldRenderReasoning } from "./reasoning.js";
+import { TerminalSpinner } from "./terminal-spinner.js";
+import { dim, supportsColor } from "./terminal-style.js";
+import { renderToolLedger, toolLedgerEntry, type ToolLedgerEntry } from "./tool-ledger.js";
+import { runClientTool } from "./tools/registry.js";
+import type { ClientToolName, ClientToolResult, ToolRunContext } from "./tools/types.js";
+import { RESUME_COMPACTION_NOTE, formatToolResultForLoop } from "./code-resume.js";
+
+
+export interface CodeAgentLoopInput {
+  task: string;
+  context: CodeContext;
+  brainUrl: string;
+  fallbackProvider?: CliProviderProfile;
+  reasoning: ReturnType<typeof parseReasoningOptions>;
+  json: boolean;
+  maxSteps: number;
+  toolCtx: ToolRunContext;
+  input: NodeJS.ReadStream;
+  output: NodeJS.WriteStream;
+  imagePaths?: string[];
+  resumeMessages?: ChatMessage[];
+  memoryFrame?: string;
+  onCheckpoint?: (line: { type: "user" | "assistant" | "tool"; content: string; data?: Record<string, unknown> }) => Promise<void>;
+  onEvent?: (event: CodeAgentEvent) => void;
+  requestApproval?: (request: CodeAgentApprovalRequest) => Promise<"approve" | "approve_all" | "deny">;
+}
+
+export interface CodeAgentApprovalRequest {
+  tool: ClientToolName;
+  args: CodeToolRequest["args"];
+  cwd: string;
+  preview?: string;
+}
+
+export type CodeAgentEvent =
+  | { type: "step.started"; step: number; label: string }
+  | { type: "provider"; provider: string }
+  | { type: "usage"; summary: string }
+  | { type: "reasoning.delta"; text: string }
+  | { type: "assistant.delta"; text: string }
+  | { type: "tool.progress"; message: string }
+  | { type: "tool.requested"; tool: ClientToolName; args: CodeToolRequest["args"]; preview?: string }
+  | { type: "tool.result"; result: ClientToolResult }
+  | { type: "tool.ledger"; text: string }
+  | { type: "tool.loop_guard"; tool: ClientToolName; repeats: number }
+  | { type: "plan.updated"; items: CodePlanItem[] }
+  | { type: "session.resumed"; path: string; messages: number }
+  | { type: "session.checkpoint"; path: string; line: "user" | "assistant" | "tool" }
+  | { type: "session.saved"; path: string }
+  | { type: "task.finished"; ok: boolean; text: string; usageSummary: string | null; maxStepsReached?: boolean; resumeCommand?: string; sessionPath?: string | null }
+  | { type: "error"; message: string };
+
+export interface CodeAgentLoopResult {
+  text: string;
+  maxStepsReached: boolean;
+  usageSummary: string | null;
+  usageRecords: Array<{ usage: unknown; durationMs: number }>;
+}
+
+interface ClientToolStormState {
+  recent: Array<{ fingerprint: string; readOnly: boolean }>;
+}
+
+interface ClientToolStormVerdict {
+  suppress: boolean;
+  repeatCount: number;
+}
+
+const TOOL_STORM_WINDOW = 8;
+const RUNTIME_COMPACTION_MAX_CHARS = 150_000;
+const RUNTIME_COMPACTION_KEEP_GROUPS = 8;
+
+function createClientToolStormState(): ClientToolStormState {
+  return { recent: [] };
+}
+
+function isReadOnlyToolRequest(request: CodeToolRequest): boolean {
+  return request.tool === "read_file" || request.tool === "grep" || request.tool === "glob";
+}
+
+function observeClientToolRequest(state: ClientToolStormState, request: CodeToolRequest): ClientToolStormVerdict {
+  if (request.tool === "update_plan") return { suppress: false, repeatCount: 1 };
+  const readOnly = isReadOnlyToolRequest(request);
+  if (!readOnly) {
+    state.recent = state.recent.filter((entry) => !entry.readOnly);
+  }
+  const fingerprint = toolRequestFingerprint(request);
+  const previous = state.recent.filter((entry) => entry.fingerprint === fingerprint).length;
+  if (previous > 0) {
+    return { suppress: true, repeatCount: previous + 1 };
+  }
+  state.recent.push({ fingerprint, readOnly });
+  if (state.recent.length > TOOL_STORM_WINDOW) {
+    state.recent.splice(0, state.recent.length - TOOL_STORM_WINDOW);
+  }
+  return { suppress: false, repeatCount: 1 };
+}
+
+export async function runCodeAgentLoop(inputData: CodeAgentLoopInput): Promise<CodeAgentLoopResult> {
+  const frames = buildCodeRuntimeFrames(inputData);
+  const initialPrompt = buildCodePrompt(inputData.task, inputData.context, inputData.imagePaths);
+  const initialContent = inputData.imagePaths?.length
+    ? await buildImagesContentParts(inputData.imagePaths, initialPrompt)
+    : initialPrompt;
+  const { messages } = buildCodeContextMessages({
+    frames,
+    resumeMessages: inputData.resumeMessages,
+    currentUserContent: initialContent,
+  });
+  const runtimeAnchorCount = messages.length;
+  let finalText = "";
+  let latestUsageSummary: string | null = null;
+  const usageRecords: Array<{ usage: unknown; durationMs: number }> = [];
+  const approvalSession = { approveAll: false };
+  const toolStorm = createClientToolStormState();
+  for (let step = 0; step < inputData.maxSteps; step += 1) {
+    const label = step === 0 ? t("spinner.coding") : t("spinner.reviewing");
+    inputData.onEvent?.({ type: "step.started", step, label });
+    const result = await collectBrainText({
+      brainUrl: inputData.brainUrl,
+      fallbackProvider: inputData.fallbackProvider,
+      messages,
+      reasoning: inputData.reasoning,
+      json: inputData.json,
+      label,
+      onEvent: inputData.onEvent,
+    });
+    const assistantText = result.text;
+    latestUsageSummary = result.usageSummary || latestUsageSummary;
+    usageRecords.push(...result.usageRecords);
+    const structuredToolRequests = toolRequestsFromCollectedCalls(result.toolCalls, step);
+    const toolRequests = structuredToolRequests.length ? structuredToolRequests : parseCodeToolRequests(assistantText);
+    messages.push(structuredToolRequests.length
+      ? {
+          role: "assistant",
+          content: assistantText,
+          tool_calls: assistantToolCallsForMessages(structuredToolRequests),
+        }
+      : { role: "assistant", content: assistantText });
+    if (inputData.onCheckpoint) {
+      if (structuredToolRequests.length) {
+        await inputData.onCheckpoint({
+          type: "assistant",
+          content: assistantText,
+          data: { tool_calls: assistantToolCallsForMessages(structuredToolRequests) },
+        });
+      } else if (assistantText.trim()) {
+        await inputData.onCheckpoint({ type: "assistant", content: assistantText });
+      }
+    }
+    if (!toolRequests.length) {
+      finalText = assistantText;
+      break;
+    }
+    const toolResultSections: string[] = [];
+    const toolLedgerEntries: ToolLedgerEntry[] = [];
+    for (const toolRequest of toolRequests) {
+      const stormVerdict = observeClientToolRequest(toolStorm, toolRequest);
+      if (inputData.json) writeJsonLine({ type: "code.tool.requested", ts: nowIso(), tool: toolRequest.tool, args: redactToolArgs(toolRequest) });
+      const preview = formatDangerousToolPreview(toolRequest.tool, toolRequest.args, supportsColor(inputData.output));
+      inputData.onEvent?.({ type: "tool.requested", tool: toolRequest.tool, args: redactToolArgs(toolRequest) as CodeToolRequest["args"], preview });
+      if (!inputData.json && !inputData.onEvent && toolRequest.tool !== "update_plan") renderClientToolStart(toolRequest);
+      let toolResult: ClientToolResult;
+      if (stormVerdict.suppress) {
+        toolResult = {
+          ok: false,
+          tool: toolRequest.tool,
+          error: "Repeated identical tool request suppressed by Lynn CLI. Use a different tool, different arguments, or answer with the information already available.",
+        };
+        if (inputData.json) {
+          writeJsonLine({ type: "code.tool.loop_guard", ts: nowIso(), tool: toolRequest.tool, args: redactToolArgs(toolRequest), repeats: stormVerdict.repeatCount });
+        }
+        inputData.onEvent?.({ type: "tool.loop_guard", tool: toolRequest.tool, repeats: stormVerdict.repeatCount });
+      } else {
+        try {
+          let effectiveApproval: ToolApprovalRequest["approval"];
+          if (
+            inputData.requestApproval &&
+            isDangerousClientTool(toolRequest.tool) &&
+            inputData.toolCtx.approval === "ask" &&
+            !approvalSession.approveAll
+          ) {
+            const decision = await inputData.requestApproval({
+              tool: toolRequest.tool,
+              args: redactToolArgs(toolRequest) as CodeToolRequest["args"],
+              cwd: inputData.toolCtx.cwd,
+              preview,
+            });
+            if (decision === "deny") throw new Error(`${toolRequest.tool} cancelled by user`);
+            if (decision === "approve_all") approvalSession.approveAll = true;
+            effectiveApproval = "yolo";
+          } else {
+            effectiveApproval = await resolveToolApproval({
+              tool: toolRequest.tool,
+              approval: inputData.toolCtx.approval,
+              cwd: inputData.toolCtx.cwd,
+              json: inputData.json,
+              input: inputData.input,
+              output: inputData.output,
+              preview,
+              session: approvalSession,
+            });
+          }
+          toolResult = await runClientTool({ ...inputData.toolCtx, approval: effectiveApproval }, {
+            name: toolRequest.tool,
+            ...toolRequest.args,
+          });
+        } catch (error) {
+          if (inputData.json && error instanceof ToolApprovalRequiredError) {
+            writeJsonLine({
+              type: "code.tool.approval_required",
+              ts: nowIso(),
+              status: "waiting_approval",
+              tool: toolRequest.tool,
+              args: redactToolArgs(toolRequest),
+              approval: inputData.toolCtx.approval,
+              sandbox: inputData.toolCtx.sandbox || "workspace-write",
+              message: error.message,
+            });
+          }
+          toolResult = {
+            ok: false,
+            tool: toolRequest.tool,
+            error: error instanceof Error ? error.message : String(error),
+          };
+        }
+      }
+      if (toolRequest.tool === "update_plan") {
+        const items = normalizePlanItems(toolRequest.args.plan);
+        if (inputData.json) writeJsonLine({ type: "code.plan.updated", ts: nowIso(), items });
+        inputData.onEvent?.({ type: "plan.updated", items });
+        if (!inputData.json && !inputData.onEvent) process.stderr.write(`${renderPlanItems(items)}\n`);
+      }
+      if (inputData.json) writeJsonLine({ type: "code.tool.result", ts: nowIso(), ...toolResult });
+      inputData.onEvent?.({ type: "tool.result", result: toolResult });
+      if (!inputData.json && !inputData.onEvent && toolRequest.tool !== "update_plan") renderClientToolResult(toolResult);
+      toolLedgerEntries.push(toolLedgerEntry(toolResult));
+      toolResultSections.push([
+        `Tool result for ${toolRequest.tool}:`,
+        formatToolResultForLoop(toolResult),
+      ].join("\n"));
+    }
+    const toolLedger = renderToolLedger(toolLedgerEntries, step);
+    if (toolLedger) {
+      if (inputData.json) writeJsonLine({ type: "code.tool.ledger", ts: nowIso(), step, text: toolLedger });
+      inputData.onEvent?.({ type: "tool.ledger", text: toolLedger });
+    }
+    if (structuredToolRequests.length) {
+      for (let i = 0; i < structuredToolRequests.length; i += 1) {
+        const request = structuredToolRequests[i];
+        const section = toolResultSections[i] || `Tool result for ${request.tool}:\n(no result captured)`;
+        messages.push({
+          role: "tool",
+          tool_call_id: request.toolCallId,
+          name: request.tool,
+          content: section,
+        });
+        if (inputData.onCheckpoint) {
+          await inputData.onCheckpoint({
+            type: "tool",
+            content: section,
+            data: {
+              tool_call_id: request.toolCallId,
+              name: request.tool,
+            },
+          });
+        }
+      }
+      if (toolLedger) {
+        const ledgerMessage = [
+          toolLedger,
+          "Continue from the ledger. If no more tools are needed, give the final answer.",
+        ].join("\n");
+        messages.push({ role: "user", content: ledgerMessage });
+        if (inputData.onCheckpoint) await inputData.onCheckpoint({ type: "user", content: ledgerMessage });
+      }
+    } else {
+      const toolResultMessage = [
+        toolResultSections.length === 1 ? "Tool results:" : `Tool results for ${toolResultSections.length} requested tools:`,
+        ...toolResultSections,
+        toolLedger,
+        "Continue. If no more tools are needed, give the final answer.",
+      ].filter(Boolean).join("\n");
+      messages.push({
+        role: "user",
+        content: toolResultMessage,
+      });
+      if (inputData.onCheckpoint) await inputData.onCheckpoint({ type: "user", content: toolResultMessage });
+    }
+    compactRuntimeMessages(messages, undefined, undefined, runtimeAnchorCount);
+  }
+  let maxStepsReached = false;
+  if (!finalText) {
+    maxStepsReached = true;
+    finalText = "Stopped after the maximum tool steps. Review the emitted tool results before continuing.";
+  }
+  return {
+    text: finalText,
+    maxStepsReached,
+    usageSummary: latestUsageSummary,
+    usageRecords,
+  };
+}
+
+export function compactRuntimeMessages(
+  messages: ChatMessage[],
+  maxChars = RUNTIME_COMPACTION_MAX_CHARS,
+  keepGroups = RUNTIME_COMPACTION_KEEP_GROUPS,
+  anchorCount = messages[0]?.role === "system" ? 1 : 0,
+): number {
+  if (messages.length < keepGroups * 2) return 0;
+  const total = messages.reduce((sum, message) => sum + runtimeMessageCost(message), 0);
+  if (total <= maxChars) return 0;
+  const prefixCount = Math.max(0, Math.min(anchorCount, messages.length));
+  const suffixGroups = buildRuntimeMessageGroups(messages.slice(prefixCount)).slice(-keepGroups);
+  const keep = suffixGroups.flat();
+  const keepSet = new Set(keep);
+  const compactable = messages.slice(prefixCount).filter((message) => !keepSet.has(message));
+  if (compactable.length < 2) return 0;
+  const summary = summarizeRuntimeMessages(compactable);
+  messages.splice(prefixCount, messages.length - prefixCount, {
+    role: "user",
+    content: `[Lynn CLI runtime compaction: ${RESUME_COMPACTION_NOTE}. Compacted ${compactable.length} older message(s) while preserving the active goal, recent tool results, and current plan. Summary:\n${summary}]`,
+  }, ...keep);
+  return compactable.length;
+}
+
+function buildRuntimeMessageGroups(turns: ChatMessage[]): ChatMessage[][] {
+  const groups: ChatMessage[][] = [];
+  for (let i = 0; i < turns.length; i += 1) {
+    const turn = turns[i];
+    if (turn.role === "assistant" && turn.tool_calls?.length) {
+      const required = new Set(turn.tool_calls.map((toolCall) => toolCall.id));
+      const group = [turn];
+      let j = i + 1;
+      while (j < turns.length && required.size > 0) {
+        const candidate = turns[j];
+        if (candidate.role !== "tool" || !candidate.tool_call_id || !required.has(candidate.tool_call_id)) break;
+        group.push(candidate);
+        required.delete(candidate.tool_call_id);
+        j += 1;
+      }
+      groups.push(group);
+      i = Math.max(i, j - 1);
+      continue;
+    }
+    if (turn.role === "tool") continue;
+    groups.push([turn]);
+  }
+  return groups;
+}
+
+function summarizeRuntimeMessages(messages: readonly ChatMessage[]): string {
+  return messages
+    .map((message, index) => {
+      const content = typeof message.content === "string" ? message.content : JSON.stringify(message.content);
+      const clean = content.replace(/\s+/g, " ").trim();
+      const label = message.role === "assistant" && message.tool_calls?.length
+        ? `assistant tool_calls=${message.tool_calls.map((toolCall) => toolCall.function.name).join(",")}`
+        : message.role;
+      return `${index + 1}. ${label}: ${clean.slice(0, 360)}${clean.length > 360 ? "..." : ""}`;
+    })
+    .join("\n")
+    .slice(0, 12_000);
+}
+
+function runtimeMessageCost(message: ChatMessage): number {
+  const contentCost = typeof message.content === "string" ? message.content.length : JSON.stringify(message.content).length;
+  const toolCallCost = message.tool_calls?.length ? JSON.stringify(message.tool_calls).length : 0;
+  return contentCost + toolCallCost;
+}
+
+interface BrainTextResult {
+  text: string;
+  usageSummary: string | null;
+  usageRecords: Array<{ usage: unknown; durationMs: number }>;
+  toolCalls: CollectedToolCall[];
+}
+
+async function collectBrainText(inputData: {
+  brainUrl: string;
+  fallbackProvider?: CliProviderProfile;
+  messages: ChatMessage[];
+  reasoning: ReturnType<typeof parseReasoningOptions>;
+  json: boolean;
+  label: string;
+  onEvent?: (event: CodeAgentEvent) => void;
+}): Promise<BrainTextResult> {
+  let text = "";
+  let usageSummary: string | null = null;
+  const usageRecords: Array<{ usage: unknown; durationMs: number }> = [];
+  const streamedToolCalls = createStreamingToolCallAccumulator();
+  const spinner = new TerminalSpinner(process.stderr, inputData.label, { quiet: true });
+  const renderState: HumanBrainRenderState = {};
+  const startedAt = Date.now();
+  if (!inputData.json && !inputData.onEvent) spinner.start();
+  try {
+    for await (const event of streamBrainChat({
+      brainUrl: inputData.brainUrl,
+      reasoning: inputData.reasoning,
+      messages: inputData.messages,
+      fallbackProvider: inputData.fallbackProvider,
+      tools: codeToolDefinitions(),
+    })) {
+      const renderReasoning = shouldRenderReasoning(inputData.reasoning.display, inputData.json);
+      if (eventWritesHumanOutput(event, renderReasoning, !!inputData.json || !!inputData.onEvent)) {
+        spinner.stop();
+      }
+      if (event.type === "reasoning.delta" && renderReasoning) {
+        inputData.onEvent?.({ type: "reasoning.delta", text: event.text });
+        if (!inputData.onEvent) {
+          if (inputData.json) writeJsonLine({ type: "reasoning.delta", ts: nowIso(), text: event.text, hidden: true });
+          else process.stderr.write(dim(event.text, supportsColor(process.stderr)));
+        }
+      }
+      if (event.type === "assistant.delta") {
+        text += event.text;
+        inputData.onEvent?.({ type: "assistant.delta", text: event.text });
+      }
+      if (event.type === "tool_call.delta") streamedToolCalls.append(event);
+      if (event.type === "provider") inputData.onEvent?.({ type: "provider", provider: event.activeProvider });
+      if (event.type === "tool_progress") inputData.onEvent?.({ type: "tool.progress", message: [event.name, event.event].filter(Boolean).join(" ") });
+      if (event.type === "usage") usageRecords.push({ usage: event.usage, durationMs: Date.now() - startedAt });
+      if (inputData.json && (event.type === "provider" || event.type === "tool_progress" || event.type === "brain.error" || event.type === "usage")) {
+        if (event.type === "usage") writeJsonLine({ type: "usage", ts: nowIso(), usage: event.usage, durationMs: Date.now() - startedAt });
+        else writeJsonLine({ ...event, ts: nowIso() });
+      }
+      if (event.type === "brain.error") {
+        throw new Error(formatBrainErrorForHuman(event.error, event.code));
+      }
+      if (!inputData.json && event.type !== "assistant.delta" && event.type !== "reasoning.delta") {
+        if (event.type === "usage") {
+          const summary = summarizeUsage(event.usage, { durationMs: Date.now() - startedAt });
+          usageSummary = summary || usageSummary;
+          if (summary) {
+            inputData.onEvent?.({ type: "usage", summary });
+            if (!inputData.onEvent) process.stderr.write(`usage: ${summary}\n`);
+          }
+        } else {
+          if (!inputData.onEvent) renderBrainEventForHuman(event, renderState, process.stderr);
+        }
+      } else if (inputData.json && event.type === "usage") {
+        usageSummary = summarizeUsage(event.usage, { durationMs: Date.now() - startedAt }) || usageSummary;
+      } else if (inputData.onEvent && event.type === "usage") {
+        const summary = summarizeUsage(event.usage, { durationMs: Date.now() - startedAt });
+        usageSummary = summary || usageSummary;
+        if (summary) inputData.onEvent({ type: "usage", summary });
+      }
+    }
+  } finally {
+    spinner.stop();
+  }
+  return { text, usageSummary, usageRecords, toolCalls: streamedToolCalls.toToolCalls() };
+}
+
+function eventWritesHumanOutput(event: BrainStreamEvent, renderReasoning: boolean, structuredOutput: boolean): boolean {
+  if (structuredOutput) return false;
+  return event.type === "assistant.delta"
+    || event.type === "provider"
+    || event.type === "tool_progress"
+    || event.type === "brain.error"
+    || event.type === "usage"
+    || (event.type === "reasoning.delta" && renderReasoning);
+}
